[Docs] Use relative links to example files
Implements relative links to the examples from the tutorial. I can’t
modify the tags from my fork, but it’s basically just copy + pasting
this to the relative tags and push em. One thing I’ve noticed, is that
the updated stuff in master doesn’t get updated in the tags, as these
are separated. Such as my last PR.

Also removed the ittalic hightlighting